Complete test for language selection component
